Updated openstack/openstack Project: openstack/cinder  23268a686dfc58f4bdc177d5b7287daf97b1a2ca
Handle OverQuota exception during volume transfer

Cinder doesn't correctly handle OverQuota during volume
transfer. It only checks quota volumes and gigabytes,
and ignores quotas in volume type. As a result, when it
exceeds volumes and gigabytes of a volume type, Cinder doesn't
raise exception and allows accept move on.

This patch is to check every over quota in overs and fixes the
problem.

Change-Id: I02ef576ecb3052724bc14a3bd2df842bec554e0d
Closes-bug: #1539459