Updated openstack/openstack Project: openstack/tripleo-heat-templates  27b58c81ce4e3141dac2f137fb9a2bca42ee88ef
Enable notifications on the overcloud

Configures all services to send notifications to rabbit.  The puppet
modules are not consistent regarding how this is done - some expose
notification config as a top-level param, others you need to set it
through a *_config structure, and cinder provides a separate class
dedicated to enabling ceilometer notifications.

Change-Id: I23e2ddad3c59a06cfbfe5d896a16e6bad2abd943