Updated openstack/openstack Project: openstack/sahara-dashboard  6cc64f7b498fe610ccb8ca9c7bc05f73e7d88be2
fix integration test screen shots

Looking at sahara dashboard as model for possibly adding
integration tests to the app-catalog.  Noticed that the
screen shots were not being saved. Adding the +e to the
script allows the screenshot copy to actually work.

The logs from patch set 7 show this working. In that patch
set I forced a screenshot even tho the test was successful.

Change-Id: Iee393a40e8a29943203cb59f6ed669db8f570d4a