Updated openstack/openstack Project: openstack/cinder  4d093d4eca67774e8b5973646b3e61552254bc53
Propoerly call oslo's Service stop and wait

We are not calling oslo_service.Service's stop method with the graceful
parameter which ensure that stop waits for threads to properly stop
instead of just forcefully stopping them.  This is not an issue for
Cinder at this moment because the rpcserver.stop call waits for
everything to gracefully stop, but we may have something there running
and we wouldn't want to kill it.

We should also call oslo_service.Service.wait method from our
Service.wait method.

This patch ensures that we are properly calling oslo_Service.Service's
methods.

Change-Id: I8764571c41509fdf1ff7950e9d8d7d6cf0d2f57d