Updated openstack/openstack Project: openstack/python-swiftclient  9460a946246437845330a8f376a51af7ce6eb4e8
Have tox manage LANG environmental var

We need to do this because some of the py34 testing that
python-swiftclient uses depends on this variable.  The reason we don't
see the issue in the gate, is because current bare-trusty images have
this set on the jenkins shell user.

When we move to just using DIBs, the variable won't be setup by
default and python3 tests will fail.

For more information:
  https://review.openstack.org/282898

Change-Id: Id9017f31b0543bccac9c07b83237b909e2bd2b0c
Signed-off-by: Paul Belanger <pabelanger@redhat.com>