Updated openstack/openstack Project: openstack/glance  971fd945340224c0f0f8bcb004ac3c7b44c57ba0
Do not use constraints for venv

The venv environment is used by infra scripts to build tarballs and
documents - and used by both check and post jobs. In post jobs, we
currently do not support constraints, so add the non-constraints install
command here so that nova-docs (publishing of documents to
docs.openstack.org) and the tarball jobs work again.

Change-Id: I7116134ca0963de001223ba905619947fcbda07a