Updated openstack/openstack Project: openstack/cinder  b0aac3b8650147e4292eb5a74eba80e576420014
Huawei: delete_snapshot need not return any value

Currently delete_snapshot return True or False to indicate
whether the snapshot is really deleted on the array. This
is useless and just write for unit tests. Now we will remove
the return value and add another unit test for delete_snapshot.

Closes-Bug: #1549584
Change-Id: I17fea61e3c0f5565ab04bc82f172e638480f45dc