Updated openstack/openstack Project: openstack/sahara  62e8caeaefdc0a27e7b313f970c2190da0cf1502
Use the integrated tempest.lib module

Tempest-lib has been reintegrated into tempest (as code, but it still provides
a stable API). The old library is still available, but tempest tests are
failing because when sahara is installed and exports the scenario tests
for python clients, python-lib is not installed (it's correcly in
test-requirements.txt, not requirements.txt), and tempest does not
install tempest-lib anymore. Hence, fix the tests to use the new
official location.

Change-Id: Iaded03ee387b4a0bbabb7bfe9e54d38e010d9c4f