Updated openstack/openstack Project: openstack/neutron  41cc3d35afa0b79e65b420cfc8bff2e00948ee24
Collect details on ARP spoof functional failures

There seems to be some race condition or corner case in the
ARP spoofing functional tests that cause them to randomly
fail in the gate but it's difficult to reproduce them
locally. This patch adds a bunch of details on failures so
we can maybe get some hints about the unexpected state of
the bridge or interfaces that is causing the failure.

Partial-Bug: #1550623
Change-Id: I15b7ab3ce2a95d2b432239d535e3700f28ad21de