Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  d8063ad5f5d15b179b4e47a645e928f690fe9809
Fix bug on missing subunit2sql data

This commit fixes an issue when handling a nonexistent subunit stream.
The script was returning on gearman an error but then continued to
use the internally queue up the subunit event to process, even though
there wasn't a stream. This would cause a stacktrace later when the
actual processing was done on the internal queue. This commit adds
the queueing inside an else block to prevent it from running when
we don't have a subunit stream.

Change-Id: I66fdc5d7ae3702411a0b42757087cf61a4c69e35