Updated openstack/openstack Project: openstack/barbican  f8384bf188e0564816b96e04c547e806039810d1
Make bandit voting as part of pep8

Run security linter bandit as part of pep8. Pep8 is the usual linter
target and thus let's use it there instead of starting another node for
this short-running job.

The job is running stable and thus this can be made voting.

A followup change will remove the non-voting bandit job.

Change-Id: I6f2cf3902a4e786a5d31a06bf66d19bc33a60b55