Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  f4667bbba45c7919259e9ce0986a226e7d9c8ada
Use first test from subunit_stream for run_at value

There is an issue when there is a backlog or a worker crashes when we
start to work through the gearman queue all the newly added runs to
the subunit2sql db have a run_at value of when it was processed, not
when it was actually run.

The subunit2sql cli (and corresponding pythong api) by default uses
the current time as the run_at time when creating a new run. However
there is option which allows you to specify the value for run_at. This
commit uses the first test run from the subunit as the run_at value to
ensure that we have a somewhat accurate run_at value for the run in
the db regardless of when the subunit stream is actually processed.

Change-Id: I55663476865ec3739faf1b077aff7a02d0e0fa79