Updated openstack/openstack Project: openstack-infra/project-config  2badcc1893adbf065b2f3b0a85e120e28c6f44c0
Add a dib-builddate file

Add a build-date timestamp file to the nodepool-base element and
output that as part of the "network-info" macro that we run at the
start of most tests.

This will allow non-priviledged users to quickly see the date the node
running their test was built on, which can be correlated to the logs
on nodepool.openstack.org to help debug issues that might have to do
with the underlying image build.

Change-Id: Id0c9f6203ed487350285031d3965bc6290370a27