Updated openstack/openstack Project: openstack/python-zaqarclient  9016dec6e0452214df399411c108a961725932cb
Update flavor on flavor create if it exists

As of now on PUT, zaqar server updates flavor
if it is already exists else it will create a new one.
However zaqar client return previous pool info if it is
already exists. This result to a confusion for operator.

The zaqar client should maitain symmetry with zaqar server.

Closes-Bug: #1532776

Change-Id: Ica8532375ed81828d09c3bdbab0fa911c155bf1a