Updated openstack/openstack Project: openstack/cinder  e787cf66f7f32446b7b18282ae0d2ab7a7ef1504
NetApp: Add Consistency Group support for E-Series

Add Consistency Group support to the E-Series driver. This
implementation utilizes the native Consistency Group feature
available on the E-Series backend to support Cinder
Consistency Groups.

CGs and standalone snapshots both utilize snapshot groups.
There is a limit of 3 snapshot groups per volume, so the number
of standalone snapshots will be limited by the number of
consistency groups that are created, and likewise the reverse.
Each CG/Snapshot Group will support up to 32 snapshots, so each
CG that a volume is a part of will reduce the number of available
standalone snapshots that can be created by 32 (from a maximum
of 96).

Implements: blueprint netapp-eseries-consistency-groups
Change-Id: Ib0fc9fa9abc6699f2971948d3d4c5e9902381072