Updated openstack/openstack Project: openstack/trove-integration  8dc1948a3a21d30b84eed7b4b42db62a7c835388
Fix kick-start for postgresql

A recent changeset (https://review.openstack.org/#/c/271096 )
broke kick-start for PostgreSQL since it forgot to add a
postgresql.conf file in the conf directory.
This has been created and corresponding medium flavors added.

The flavors created by redstack were also changed to facilitate
testing of non-MySQL datastores in that changeset.
Using string IDs was removed from it when it was discovered
that it cause the kilo tests to fail.

This changeset puts them back (with proper handling for kilo) and
consolidates all the creation in one spot.

The vertica.conf file was also fixed to use the proper resize flavor.

Closes-Bug: #1549604
Change-Id: I326080f45a119c0e377498a7e9dfb1c8f39c5dbb