Updated openstack/openstack Project: openstack/nova  5e9b023dd7a24f988f536da9ee2d1a788ca19a78
Add some obvious detail to nw_info warning log

When we log a warning due to a missing nw_info cache we do not mention
that this log is occuring due to a missing nw_info cache. Add this
detail to make debugging easier for someone unfamiliar with this
codepath.

Change-Id: I33ec36398e95fc060e28faad86ed4c65007d6f26