Updated openstack/openstack Project: openstack-infra/tripleo-ci  11974ca57994da6e7e02d48764743507eb3fa188
Convert the container job to a noop

This job has been failing for over a month, timing out after
nearly 3 hours. As a result it is wasting 1/4 of our resources and
delaying ci reporting back to developers by up to an hour for all
patches.

Cause it to fail fast until it is fixed. We could remove it altogether
but this would make it more difficult to test when trying to get it
back up and running.

Change-Id: I45209839a2fbde0970a7f9f6f7fb89f6b1bcdc2f