Updated openstack/openstack Project: openstack/nova  3087672083a05a28df1073945c3e675a26459e4b
Fix Cells RPC API by accepting a RequestSpec arg

Since cells v1 adds a proxy to a call to ComputeTaskAPI for live_migrate
and rebuild_instance, we need to augment the Cells RPC API to accept
the request_spec argument for both methods.

That said, since cells v1 is quite feature-freeze, it's okay to not pass
the RequestSpec down to the pipe because the cell conductor will then
act as if the instance was not having a Spec object, which is a compat
behaviour we want to keep for a very long time.

Closes-Bug: #1550018

Change-Id: I1c2aaedfc83e4e97f53c9c504e4585b1f0510d4b