Updated openstack/openstack Project: openstack/cinder  1667891d31faee51ce37310a67f029dc3f7e276e
EMC VMAX - get iscsi ip from port in existing MV

This patch fixes bug in getting the iscsi ip address from
an existing masking view.  In this scenario we must query
the port(s) in the portgroup belonging to the masking view
and not the random port group from the xml file.

Change-Id: I4f8bea5c9f8c07a793aa1053678a7a915d565d22
Closes-Bug: #1504460