Updated openstack/openstack Project: openstack/neutron  afe1a834000d33900b8646d308fa26fa807a2ca0
Block delete_(network|subnet) transactioned calls

These functions assume that they won't be called inside of a
transaction for two reasons. The first is an infinite loop that
will only terminate if it can actually retrieve the latest
information from the database. The second is that ML2's post-commit
operations are expected to occur after data was actually committed
so mechanism drivers that do DB lookups using a different session
may break.

This adds an explicit guard that prevents these two functions from
being called within an active DB session. The rest of the ML2
functions should eventually be converted to this style as well
(see the partial bug).

Related-Bug: #1551958
Partial-Bug: #1540844
Change-Id: I5c00b186585369ef6c8e2b9cb5a43b8bba0e5a7c