Updated openstack/openstack Project: openstack/zaqar  f35742115f1c10745246b155e9ea1ab94797e7d2
"group" word should be used instead "pool"

Currently, when creating a new flavor, here
it need to set pool group name for the first
argument "pool".
For making clear between a pool and pool group,
So we should replace "pool" word with "pool_group"
at all the places.

According the discussion in IRC, for giving users time
to mirgate, we will support 'pool' and 'pool_group'
both in M, and will remove 'pool' in N.

APIImpact

Change-Id: Id4c976a1fc9cc7fb9b72482b4e3a7f7f80f94492
Closes-Bug: #1504959