Updated openstack/openstack Project: openstack/keystone  1ee14754ea1581fa57f25dfae0868c00b9bef728
Fix the migration issue for the user doesn't have a password

If user doesn't have a password, the migration 091 will fail, this is
because the `password` table doesn't allow the `password`
field to be null.

There is no need to insert an empty entry into `password`
table. The user not reference with any `password` instance will imply
the `user` doesn't have a password.

This patch also fixes the case that the user from `user` table already
exists in `local_user` table and result in `DBDuplicateEntry` exception
due to the bug #1549705.

Change-Id: I4db73ee673573ab63b36bf092c17f8ba64605939
Closes-Bug: #1549705