Updated openstack/openstack Project: openstack/cinder  dfe38e3fa09a809a00e8f57d33122b530f0ebaf6
Remove unused columns from Service ORM model

rpc_available_version and object_available_version columns were part of
previous design of our RPC API compatibility layer. We've decided to go
in a little different direction so these two are currently unused.

With our live-schema-upgrades guidelines (related to rolling upgrades
support) it's a lot easier to readd something to the DB than to drop it.
Let's take the chance to actually drop these unused DB columns in early
Newton. To do that we need to remove them from the SQLAlchemy model in
Mitaka and this commit does so.

Change-Id: I3b7588214863ed008a0783167580953c22f58e99
Related-Blueprint: rpc-object-compatibility