Updated openstack/openstack Project: openstack/glance  7d4e8be47348f21d5f9b622e4980bdb2fcd84345
Allow mutable argument to be passed to BinaryObject artifacts

According to spec description [1] binary data should always be
immutable. Before this commit, passing 'mutable' argument lead
to exception, because it was passed to a function twice. This commit
makes BinaryObject's silently ignore mutable=False and raises an
appropriate exception in case mutable=True is passed.

Change-Id: Ib42871004c792b840f8af0316c3b8177c4ecf706
Closes-Bug: #1542430