Updated openstack/openstack Project: openstack/nova  5cea66336ecf6822dd6c3a2c81579e2671f0e0d6
libvirt: Add ppc to supported arch for NUMA

This patch includes the ppc64/ppc64le architectures to the list
of host architectures that support NUMA via the libvirt driver.

This change is for the PowerPC based servers that use the libvirt
driver. This also introduces the minimum libvirt version needed
for PowerPC to support NUMA from Nova.

Also note, that in order to retain the unit test that excluded
PPC from the NUMA supported architectures, an arbitrary choice
of architecture is made (S390) - that is not a part of the list
that supports NUMA.

Change-Id: I51dd3e160153b40072a04095e9bf3f15abd74252
Closes-Bug: #1516578