Updated openstack/openstack Project: openstack/nova  baf8f80848cbe45c0f9f00db7d96c5c165049e6f
complete the removal of api_version from rest client parameters

This completes the removal of the api_version parameter from
individual rest calls. This is an attribute of the api rest client
itself, and should be set there and have all requests work on that
version.

In order to do that test_server_group had to be pretty substantially
normalized. This set of tests so aggressively used common methods
(including embedded asserts) that assumed a very specific execution
environment and didn't account for other microversion changes. This
expands this back out via copy/paste to decouple these components that
do vary.

Change-Id: I06b0087b755b0d19fb51ef0463d36dd2cac5f044