Updated openstack/openstack Project: openstack/keystone  0b18edab226f6e9dc531febd4eb6f65ccd3c031e
Race condition in keystone domain config

This bug fixes a race condition in the domains_config
decorator. The race condition occurs when more than
one thread accesses the decorator. The first thread
sets the configured flag to True before proceeding with
driver load leading the second thread to use the default
driver. This fix ensures that the second thread waits for
the first thread to finish configuration before it uses
the driver.

Change-Id: I0289a4d38e0d30d39c67e29bf77b0a89d1dd23f6
Closes-Bug: 1549726