Updated openstack/openstack Project: openstack/nova  67d9dad6aba0f5f92140a272fba18c7a7323331c
libvirt: set libvirt.sysinfo_serial='none' for virt driver tests

The default value for the sysinfo_serial option is 'auto' which
tries to read from the /etc/machine-id file if it exists and is
not empty.

Unit tests should not be dependent on trying to read from system
files outside the virtual environment, so for the generic virt
driver tests set the config option value to 'none' so we don't
do anything for the sysinfo_serial option.

Change-Id: Ib9ba0ec2c1690d93eadb63e699740e4ba988cce1
Closes-Bug: #1551900