Updated openstack/openstack Project: openstack/glance_store  fb77cb73c5daa9f78dbf13d9c943c91f92ba0298
swift: Do not search storage_url for ks v2

Previously, we introduced a re-auth approach for swift store.
In case of single-tenant store we stated that it works with
Keystone v3 only when re-authenticating but we also need to make
the same restriction when receiving swift endpoint from keystone
becase it breaks some installations where auth_url contains
trailing slash.

So now if auth_version is not v3 then we don't search endpoint url
from Keystone in this case and simply return None. It is safe
because we don't do any re-authentication for v1 or v2 and use
old methods for that version.

Change-Id: Id8dab9ed74eef56ffa4937bf29f96888b673ad64
Closes-Bug: #1552132