Updated openstack/openstack Project: openstack/oslo.messaging  363c591ffb339dd5ea5f935a4c9649b64ce87183
Fail quickly if there on bad password

Inspired by:
https://bugs.launchpad.net/oslo.messaging/+bug/1508512/comments/1

Add authentication_failure_close capability on the client to
tell the server that it can send us connection.close command
indicating ACCESS_REFUSED as the reasona. If this capability is
absent then authentication failures are reported in the legacy
fashion: by abruptly closing the network connection.

Partial-Bug: #1508512
Change-Id: Icf6108678256e6f922620837c71f8827fe8e8c52