Updated openstack/openstack Project: openstack-infra/project-config  397caee2a71574bf6a85f75a5efddfeeb13bcd85
Skip long-running jobs on neutronclient doc changes

Change [1] revealed that we are running tempest or unit tests
on doc changes when there is no need to. This change introduces
a skip-if clause to save nodes and execution time for changes
that only need to be validated on pep8/docs/releasenotes jobs.

[1] https://review.openstack.org/#/c/282555/

Change-Id: I5467472d82ddcf5dc3be939458840c3c0423526e