Updated openstack/openstack Project: openstack/heat  1211d6340e01b4b75dd3fb4eed2d256232ed2e36
Build list of enabled os-collect-config collectors

When changing the software_config_transport for an existing server,
os-collect-config will continue using the data from the old transport
which will become increasingly stale and will sometimes mask values from
the new transport.

The default list of collectors in os-collect-config is:
heat_local, ec2, cfn, heat, request, local, zaqar

This change explicitly enables these collectors:
ec2, <current transport>, local

``ec2`` is the nova metadata service, so should always be enabled.

``local`` allows local file overrides - this is generally unused but is
enabled for users who choose to use it for additional config data, or to
override the heat provided data.

``heat_local`` is only used on the first run of an unconfigured
os-collect-config for bootstrapping its own config file, so it can and
should be disabled for subsequent polling.

Change-Id: I9cdb057212c7a4404f6966a55cde9eab5ad86759
Closes-Bug: #1523676